+++ sso/tests/management/commands/test_archive_users.py
@@ -1,31 +1,8 @@
-from datetime import datetime, timedelta
 from unittest.mock import patch
 
 import pytest
 from django.contrib.auth import get_user_model
 from django.core.management import call_command
-
-
-@pytest.fixture
-def user():
-    User = get_user_model()  # noqa
-    user = User.objects.create(email='abc@xyz.com')
-    yield user
-    user.delete()
-
-
-@pytest.fixture
-def old_user():
-    """Fixture for old user who's last login and created three years ago"""
-    User = get_user_model()  # noqa
-    user = User.objects.create(email='def@xyz.com')
-    # user last_login around 4 years ago
-    user.last_login = datetime.now() - timedelta(days=4 * 365)
-    # user created 5 years ago
-    user.created = datetime.now() - timedelta(days=5 * 365)
-    user.save()
-    yield user
-    user.delete()
 
 
 class MockResponse:
@@ -37,7 +14,7 @@ class MockInvalidResponse:
 
 
 @pytest.mark.django_db
-def test_archive_command_users(user, old_user):
+def test_archive_command_users(active_user, old_user):
     User = get_user_model()  # noqa
     total_users = User.objects.count()
 
@@ -56,7 +33,27 @@ def test_archive_command_users(user, old_user):
 
 
 @pytest.mark.django_db
-def test_archive_command_for_recent_users(user):
+def test_archive_command_users_for_old_user_without_notification(active_user, old_user_with_no_notification):
+    User = get_user_model()  # noqa
+    total_users = User.objects.count()
+
+    assert total_users == 2
+
+    # no user should be archived as no final notification for this user
+    with patch('directory_api_client.company.CompanyAPIClient.delete_company_by_sso_id') as mock_call:
+        mock_call.return_value = MockResponse
+        # one old user deleted as per data retention policy
+        call_command('archive_users')
+
+    total_users = User.objects.count()
+
+    assert total_users == 2
+    assert mock_call.called is False
+    assert mock_call.call_count == 0
+
+
+@pytest.mark.django_db
+def test_archive_command_for_recent_users(active_user):
     User = get_user_model()  # noqa
     total_users = User.objects.count()
 
@@ -75,7 +72,7 @@ def test_archive_command_for_recent_users(user):
 
 
 @pytest.mark.django_db
-def test_archive_command_for_invalid_api_response(user, old_user):
+def test_archive_command_for_invalid_api_response(active_user, old_user):
     User = get_user_model()  # noqa
     total_users = User.objects.count()